requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
psutil>=5.9.0
bcrypt>=4.1.0
python-jose>=3.3.0
